Adjusted Upgrades and Particle Effects
- Improved Upgrade Screen UI
- Implemented Countdown Timer when Damage Upgrade is activated
- "Impact Effect" is now used when an enemy is hit so the player knows when damage is done

+++ COP 4331 - Space Shooter 3D/Assets/Scripts/Bullet.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class Bullet : MonoBehaviour
 {
-	public float damage = 0;
+	public double damage = 0;
 	public float speed = 0;
 	public GameObject player;
 	private Vector3 forward;
