+++ COP 4331 - Space Shooter 3D/Assets/Scripts/EnemyScripts/EnemyBullet.cs
@@ -2,7 +2,7 @@ using UnityEngine;
 
 public class EnemyBullet : MonoBehaviour
 {
-	public float damage = 30f;
+	public double damage = 30f;
 	public float deathTime = 16f;
 	void Start()
 	{
